Extract basic participant filter form
