+++ .idea/inspectionProfiles/Project_Default.xml
@@ -15,7 +15,7 @@
     <inspection_tool class="HtmlUnknownTag" enabled="true" level="WARNING" enabled_by_default="true">
       <option name="myValues">
         <value>
-          <list size="8">
+          <list size="9">
             <item index="0" class="java.lang.String" itemvalue="nobr" />
             <item index="1" class="java.lang.String" itemvalue="noembed" />
             <item index="2" class="java.lang.String" itemvalue="comment" />
@@ -24,6 +24,7 @@
             <item index="5" class="java.lang.String" itemvalue="script" />
             <item index="6" class="java.lang.String" itemvalue="wicket:message" />
             <item index="7" class="java.lang.String" itemvalue="wicket:panel" />
+            <item index="8" class="java.lang.String" itemvalue="wicket:participantfilterform" />
           </list>
         </value>
       </option>
